vfs: readlink() takes bufsize as size_t
This should probably done consistently for the rest of these functions,
but readlink is the only one right now that doesn't match its syscall
definition, so starting there.  Enforce that the passed in size can be
represented by a signed int before passing to the FS code (which still
uses int).

+++ vfs/vfs.h
@@ -154,7 +154,7 @@ int vfs_fchmod(int fd, kmode_t mode);
 int vfs_symlink(const char* target, const char* link);
 
 // Read the contents of a symlink into the given buffer.
-int vfs_readlink(const char* path, char* buf, int bufsize);
+int vfs_readlink(const char* path, char* buf, size_t bufsize);
 
 // Checks if the current [real] user ID has access to the given file.  Returns 0
 // on success, or -error.
